Extract shared_examples for propagation
