+++ spec/datadog/tracing/contrib/mysql2/patcher_spec.rb
@@ -4,6 +4,7 @@ require 'datadog/tracing/contrib/integration_examples'
 require 'datadog/tracing/contrib/support/spec_helper'
 require 'datadog/tracing/contrib/analytics_examples'
 require 'datadog/tracing/contrib/propagation/sql_comment'
+require 'datadog/tracing/contrib/sql_comment_propagation_examples'
 
 require 'ddtrace'
 require 'mysql2'
@@ -109,33 +110,6 @@ RSpec.describe 'Mysql2::Client patcher' do
         let(:sql_statement) { 'SELECT 1' }
 
         subject { client.query(sql_statement) }
-
-        shared_examples_for 'propagated with sql comment propagation' do |mode, span_op_name|
-          it "propagates with mode: #{mode}" do
-            expect(Datadog::Tracing::Contrib::Propagation::SqlComment::Mode)
-              .to receive(:new).with(mode).and_return(propagation_mode)
-
-            subject
-          end
-
-          it 'decorates the span operation' do
-            expect(Datadog::Tracing::Contrib::Propagation::SqlComment).to receive(:annotate!).with(
-              a_span_operation_with(name: span_op_name),
-              propagation_mode
-            )
-            subject
-          end
-
-          it 'prepends sql comment to the sql statement' do
-            expect(Datadog::Tracing::Contrib::Propagation::SqlComment).to receive(:prepend_comment).with(
-              sql_statement,
-              a_span_operation_with(name: span_op_name, service: service_name),
-              propagation_mode
-            ).and_call_original
-
-            subject
-          end
-        end
 
         context 'when default `disabled`' do
           it_behaves_like 'propagated with sql comment propagation', 'disabled', 'mysql2.query' do
